+++ src/main/java/org/opensearch/neuralsearch/processor/normalization/ScoreNormalizationFactory.java
@@ -4,7 +4,14 @@
  */
 package org.opensearch.neuralsearch.processor.normalization;
 
+import org.opensearch.neuralsearch.processor.TechniqueCompatibilityCheckDTO;
+import org.opensearch.neuralsearch.processor.combination.ArithmeticMeanScoreCombinationTechnique;
+import org.opensearch.neuralsearch.processor.combination.GeometricMeanScoreCombinationTechnique;
+import org.opensearch.neuralsearch.processor.combination.HarmonicMeanScoreCombinationTechnique;
+
+import java.util.Locale;
 import java.util.Map;
+import java.util.Set;
 import java.util.Optional;
 import java.util.function.Function;
 
@@ -17,13 +24,38 @@ public class ScoreNormalizationFactory {
 
     public static final ScoreNormalizationTechnique DEFAULT_METHOD = new MinMaxScoreNormalizationTechnique();
 
-    private final Map<String, Function<Map<String, Object>, ScoreNormalizationTechnique>> scoreNormalizationMethodsMap = Map.of(
+    private static final Map<String, Function<Map<String, Object>, ScoreNormalizationTechnique>> SCORE_NORMALIZATION_METHODS = Map.of(
         MinMaxScoreNormalizationTechnique.TECHNIQUE_NAME,
         params -> new MinMaxScoreNormalizationTechnique(params, scoreNormalizationUtil),
         L2ScoreNormalizationTechnique.TECHNIQUE_NAME,
         params -> new L2ScoreNormalizationTechnique(params, scoreNormalizationUtil),
         RRFNormalizationTechnique.TECHNIQUE_NAME,
-        params -> new RRFNormalizationTechnique(params, scoreNormalizationUtil)
+        params -> new RRFNormalizationTechnique(params, scoreNormalizationUtil),
+        ZScoreNormalizationTechnique.TECHNIQUE_NAME,
+        params -> new ZScoreNormalizationTechnique()
+    );
+
+    private static final Map<String, Set<String>> COMBINATION_TECHNIQUE_FOR_NORMALIZATION_METHODS = Map.of(
+        MinMaxScoreNormalizationTechnique.TECHNIQUE_NAME,
+        Set.of(
+            ArithmeticMeanScoreCombinationTechnique.TECHNIQUE_NAME,
+            GeometricMeanScoreCombinationTechnique.TECHNIQUE_NAME,
+            HarmonicMeanScoreCombinationTechnique.TECHNIQUE_NAME
+        ),
+        L2ScoreNormalizationTechnique.TECHNIQUE_NAME,
+        Set.of(
+            ArithmeticMeanScoreCombinationTechnique.TECHNIQUE_NAME,
+            GeometricMeanScoreCombinationTechnique.TECHNIQUE_NAME,
+            HarmonicMeanScoreCombinationTechnique.TECHNIQUE_NAME
+        ),
+        RRFNormalizationTechnique.TECHNIQUE_NAME,
+        Set.of(
+            ArithmeticMeanScoreCombinationTechnique.TECHNIQUE_NAME,
+            GeometricMeanScoreCombinationTechnique.TECHNIQUE_NAME,
+            HarmonicMeanScoreCombinationTechnique.TECHNIQUE_NAME
+        ),
+        ZScoreNormalizationTechnique.TECHNIQUE_NAME,
+        Set.of(ArithmeticMeanScoreCombinationTechnique.TECHNIQUE_NAME)
     );
 
     /**
@@ -36,8 +68,30 @@ public class ScoreNormalizationFactory {
     }
 
     public ScoreNormalizationTechnique createNormalization(final String technique, final Map<String, Object> params) {
-        return Optional.ofNullable(scoreNormalizationMethodsMap.get(technique))
+        return Optional.ofNullable(SCORE_NORMALIZATION_METHODS.get(technique))
             .orElseThrow(() -> new IllegalArgumentException("provided normalization technique is not supported"))
             .apply(params);
     }
+
+    /**
+     * Validate normalization technique based on combination technique and other params that needs to be validated
+     * @param techniqueCompatibilityCheckDTO data transfer object that contains combination technique and other params that needs to be validated
+     */
+    public void isTechniquesCompatible(TechniqueCompatibilityCheckDTO techniqueCompatibilityCheckDTO) {
+        ScoreNormalizationTechnique normalizationTechnique = techniqueCompatibilityCheckDTO.getScoreNormalizationTechnique();
+        Set<String> supportedTechniques = COMBINATION_TECHNIQUE_FOR_NORMALIZATION_METHODS.get(normalizationTechnique.techniqueName());
+
+        if (supportedTechniques.contains(techniqueCompatibilityCheckDTO.getScoreCombinationTechnique().techniqueName()) == false) {
+            throw new IllegalArgumentException(
+                String.format(
+                    Locale.ROOT,
+                    "provided combination technique %s is not supported for normalization technique %s. Supported techniques are: %s",
+                    techniqueCompatibilityCheckDTO.getScoreCombinationTechnique().techniqueName(),
+                    normalizationTechnique.techniqueName(),
+                    String.join(", ", supportedTechniques)
+                )
+            );
+        }
+    }
+
 }